I've added a simple benchmark based on @jkotas sample (https://github.com/jkotas/commandline-perf/blob/master/command-line-api/Program.cs)

`String[0]` is test case when user has not passed any args
`String[4]` is for `--bool true -s test`

| Method |      Args | Mean before | Mean after |
|------- |---------- |------------:|-----------:|
|   Sync | String[0] |    88.04 us |   45.60 us |
|  Async | String[0] |    61.21 us |   46.05 us |
|   Sync | String[4] |    85.71 us |   50.65 us |
|  Async | String[4] |    66.33 us |   51.73 us |

Explanation:
* there is no need to start a new task, we can call `invocationChain` (it returns a task) and wait for it directly
* there is no need to check if directory exists if we know that the file that belongs to this directory exists

@jonsequitur PTAL