@jkotas 2021-12-02
This is good change. It would be also interesting to measure the total time to execute the program w/ the command line parser (ie include JITing and type loading). It is where the most severe System.CommandLine perf issues has been.