@jonsequitur 2021-12-02
@adamsitnik Please also feel free to prune the other benchmarks.  They're maybe more granular than necessary.